modules: openthread: Split milli and micro second implementation
Split the milli/micro second APIs in OpenThread alarm implementation
so that it is easier to have multiple driver implementations
for them.
